Server side handling of Init and KeepAlive message
Basic Server Side handling of `init` and `keepalive` messages is
complete. We verify that the `agentUUID` received in the message, indeed
identifies an agent or else return a 404 not found.

For keepalive message, for each of the 'services' received in keepalive
message - we mark the service as "Active" (create new `PktTrailService`
if the service object already does not exist.) The missing ones from
latest keepalive message are marked as inactive.

We are also sending a proper response using the schema, so this can be
validated at the client side.