Using `method_to_schema_class` from `pkttrail.schema`
`pkttrail.schema.messages` exported a dict called
`method_to_schema_class` to get the required `Request`/`Response`
schema. Initially we were using a local 'private' dictionary for the
same, however it's a good idea that all clients use it from a single
source of truth that is the 'plttrail.schema' sub-package.