Resource State book-keeping through celery task(s)
A resource state is 'online' when we hear from the agent running on the
resource. However, if we do not hear from the agent for certain time
('n' Keep Alive messages lost), we have to mark the resource as offline.
This is handled through a celery beat periodic task, that periodically
goes through a list of all resources which are 'online'. If we've not
heard for KeepAlive timeout seconds from them, we mark those resources
as offline and log.